requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20200923.3=h9c3ff4c_0
absl-py=0.13.0=py39h06a4308_0
aiohttp=3.7.4=py39h27cfd23_1
alembic=1.6.5=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
asn1crypto=1.4.0=pyh9f0ad1d_0
astor=0.8.1=py39h06a4308_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py39h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
bottleneck=1.3.2=py39hdd57654_1
brotlipy=0.7.0=py39h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py39h06a4308_0
cffi=1.14.5=py39h261ae71_0
chardet=3.0.4=py39h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
configparser=5.0.2=pyhd8ed1ab_0
coverage=5.5=py39h27cfd23_2
cryptography=3.4.7=py39hd23ed53_0
cycler=0.10.0=py39h06a4308_0
cython=0.29.23=py39h2531618_0
databricks-cli=0.12.1=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd8ed1ab_0
dm-tree=0.1.5=py39hde0f152_2
docker-py=5.0.0=py39hf3d152e_0
docker-pycreds=0.4.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h2531618_2
flask=1.1.2=pyh9f0ad1d_0
flatbuffers=2.0.0=h9c3ff4c_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
gast=0.4.0=py_0
giflib=5.2.1=h36c2ea0_2
gitdb=4.0.7=pyhd8ed1ab_0
gitpython=3.1.18=pyhd8ed1ab_0
glib=2.69.0=h5202010_0
google-auth=1.32.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
gorilla=0.4.0=pyhd8ed1ab_0
greenlet=1.1.0=py39he80948d_0
grpcio=1.36.1=py39h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gunicorn=20.1.0=py39hf3d152e_0
h5py=3.2.1=py39h6c542dc_0
hdf5=1.10.6=hb1b8bf9_0
icu=68.1=h58526e2_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py39h06a4308_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py39hb070fc8_0
ipython=7.22.0=py39hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
itsdangerous=2.0.1=pyhd8ed1ab_0
jedi=0.17.2=py39h06a4308_1
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py39h06a4308_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py39h2531618_0
krb5=1.19.1=hcc1bbae_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libcurl=7.77.0=h2574ce0_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h516909a_0
libnghttp2=1.43.0=h812cca2_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h72842e0_0
lz4-c=1.9.3=h2531618_0
mako=1.1.4=pyh44b312d_0
markdown=3.3.4=py39h06a4308_0
markupsafe=2.0.1=py39h3811e60_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py39h98787fa_1
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py39h27cfd23_1
mkl_fft=1.3.0=py39h42c9631_2
mkl_random=1.2.1=py39ha9443f7_2
mlflow=1.2.0=py_1
multidict=5.1.0=py39h27cfd23_2
ncurses=6.2=he6710b0_1
numexpr=2.7.3=py39h22e1b3c_1
numpy=1.20.2=py39h2d18471_0
numpy-base=1.20.2=py39hfae3a4d_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.0=pyhd8ed1ab_0
pandas=1.2.5=py39h295c915_0
parso=0.7.0=py_0
patsy=0.5.1=py39h06a4308_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py39h2c7a002_0
pip=21.1.3=py39h06a4308_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.14.0=py39h2531618_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pydantic=1.8.2=py39h27cfd23_0
pygments=2.9.0=pyhd3eb1b0_0
pyjwt=2.1.0=py39h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.5=h12debd9_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-editor=1.0.4=py_0
python-flatbuffers=1.12=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py39h3811e60_0
pyzmq=20.0.0=py39h2531618_1
querystring_parser=1.2.4=py_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.2=py39ha9443f7_0
scipy=1.6.2=py39had2a1c9_1
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py39h06a4308_0
simplejson=3.17.2=py39h3811e60_2
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.8=he1b5a44_3
sqlalchemy=1.4.20=py39h3811e60_0
sqlite=3.36.0=hc218d9a_0
sqlparse=0.4.1=pyh9f0ad1d_0
statsmodels=0.12.2=py39h27cfd23_0
tabulate=0.8.9=py39h06a4308_0
tenacity=8.0.0=pyhd8ed1ab_0
tensorboard=2.5.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39h3da14fd_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.5.0=mkl_py39h4a0693c_0
tensorflow-base=2.5.0=mkl_py39h35b2a3d_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
tensorflow-probability=0.13.0=pyhd8ed1ab_0
termcolor=1.1.0=py39h06a4308_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.1=py39h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=py_0
websocket-client=0.57.0=py39hf3d152e_4
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.35.1=pyh9f0ad1d_0
wrapt=1.12.1=py39he8ac12f_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py39h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
